requirements.txt
openstacksdk>=2.0.0
python-keystoneclient  # Required for Rackspace authentication
